+++ lib/libc.dart
@@ -3,9 +3,7 @@ library libc;
 import "dart:io";
 
 import "package:binary_interop/binary_interop.dart";
-import "package:binary_marshalling/binary_resource.dart";
 import "package:binary_types/binary_types.dart";
 
-part "src/libc/libc_generated.dart";
+part "src/libc/libc_native.dart";
 part "src/libc/libc_library.dart";
-part "src/libc/libc_managed.dart";