+++ lib/src/headers/headers.dart
@@ -390,6 +390,10 @@ const String _WCHAR_H = """
 
 #include <stddef.h>
 
+typedef wchar_t wint_t;
+// mbstate_t
+// wctype_t
+
 #endif
 """;